perf(dashboard): code-split chart-options and defer slot-driven recomputes (P1)
P0 で抑えた再レンダの次に、初回バンドル・スライダ応答性・レンダラ選定を最適化する。

- chart-options.ts (1,025行) を機能別に分割:
  - chart-options/shared.ts (helpers)
  - chart-options/reserves.ts (予備率/需要/予備力スナップショット)
  - chart-options/generation.ts (発電トレンド/構成比/エリア集計)
  - chart-options/flows.ts (エリア間潮流/連系線トレンド/ヒートマップ)
  - chart-options/generator-status.ts (treemap + per-area time series)
  - chart-options/congestion.ts (旧 chart-options-congestion.ts を移設)
  - chart-options/index.ts (barrel)
  dashboard-app.tsx 等は機能モジュールから直接 import し、tree-shaking を効かせる
- 条件付き表示の重いセクションを next/dynamic で遅延ロード:
  CongestionSection / JepxMarketCard / JepxAreaBreakdown / GeneratorStatusSection
- エリア別需給カードのスパークラインを SVG → Canvas レンダラへ切替
- 時刻スライダ操作中の応答性向上: useDeferredValue でチャート系の重い計算を
  非同期化。スライダ thumb と時刻ラベルは即時反映、チャート更新は遅延追従
