+++ src/lib/chart-options/index.ts
@@ -1,0 +1,14 @@
+/**
+ * ECharts option builders — barrel re-export.
+ *
+ * Prefer importing from the specific feature module (e.g.
+ * `@/lib/chart-options/reserves`) when only a subset is needed so bundlers
+ * can tree-shake more effectively.
+ */
+
+export * from "./shared";
+export * from "./reserves";
+export * from "./generation";
+export * from "./flows";
+export * from "./generator-status";
+export * from "./congestion";
